chore: move tests into enterprise
