+++ cli/organizationmembers_test.go
@@ -9,7 +9,6 @@ import (
 	"github.com/coder/coder/v2/cli/clitest"
 	"github.com/coder/coder/v2/coderd/coderdtest"
 	"github.com/coder/coder/v2/coderd/rbac"
-	"github.com/coder/coder/v2/codersdk"
 	"github.com/coder/coder/v2/testutil"
 )
 
@@ -36,42 +35,7 @@ func TestListOrganizationMembers(t *testing.T) {
 	})
 }
 
-func TestAddOrganizationMembers(t *testing.T) {
-	t.Parallel()
 
-	t.Run("OK", func(t *testing.T) {
-		t.Parallel()
-
-		ownerClient := coderdtest.New(t, &coderdtest.Options{})
-		owner := coderdtest.CreateFirstUser(t, ownerClient)
-		_, user := coderdtest.CreateAnotherUser(t, ownerClient, owner.OrganizationID)
-
-		ctx := testutil.Context(t, testutil.WaitMedium)
-		//nolint:gocritic // must be an owner, only owners can create orgs
-		otherOrg, err := ownerClient.CreateOrganization(ctx, codersdk.CreateOrganizationRequest{
-			Name:        "Other",
-			DisplayName: "",
-			Description: "",
-			Icon:        "",
-		})
-		require.NoError(t, err, "create another organization")
-
-		inv, root := clitest.New(t, "organization", "members", "add", "-O", otherOrg.ID.String(), user.Username)
-		//nolint:gocritic // must be an owner
-		clitest.SetupConfig(t, ownerClient, root)
-
-		buf := new(bytes.Buffer)
-		inv.Stdout = buf
-		err = inv.WithContext(ctx).Run()
-		require.NoError(t, err)
-
-		//nolint:gocritic // must be an owner
-		members, err := ownerClient.OrganizationMembers(ctx, otherOrg.ID)
-		require.NoError(t, err)
-
-		require.Len(t, members, 2)
-	})
-}
 
 func TestRemoveOrganizationMembers(t *testing.T) {
 	t.Parallel()
