+++ enterprise/coderd/templates_test.go
@@ -1847,6 +1847,7 @@ func TestMultipleOrganizationTemplates(t *testing.T) {
 		LicenseOptions: &coderdenttest.LicenseOptions{
 			Features: license.Features{
 				codersdk.FeatureExternalProvisionerDaemons: 1,
+				codersdk.FeatureMultipleOrganizations:      1,
 			},
 		},
 	})
